Simplify 'invalid cashes' verification
Instead of checking cachify prefix against the md5 hash of the file
content we just check if the said prefix is among the ones that we
previously generated.

It provides a safeguard against someone maliciously attempting to clog
the downstream cache by sending us a great number of invalid URLs and
has an added benefit of beeing much simpler than 'recalculate the hash'
method.

Fixes #37